pdf-security-app_04-mini-high_30-04-25/analysis_service/requirements.txt
flask
PyPDF2
pdfminer.six
python-json-logger
